requirements.txt
pandas>=1.0.0
jupyter
matplotlib
seaborn
plotly
python-docx
pycountry
kaleido
